Move ctype.h to right location
