+++ library/src/ctype/isalnum.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isalnum(int c)
 {
