+++ library/src/ctype/isalpha.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isalpha(int c)
 {
