+++ library/src/ctype/isblank.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isblank(int c)
 {
