+++ library/src/ctype/iscntrl.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int iscntrl(int c)
 {
