+++ library/src/ctype/isdigit.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isdigit(int c)
 {
