+++ library/src/ctype/isgraph.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isgraph(int c)
 {
