+++ library/src/ctype/islower.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int islower(int c)
 {
