+++ library/src/ctype/isprint.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isprint(int c)
 {
