+++ library/src/ctype/isupper.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isupper(int c)
 {
