+++ library/src/ctype/isxdigit.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int isxdigit(int c)
 {
