+++ library/src/ctype/tolower.c
@@ -1,4 +1,4 @@
-#include "ctype.h"
+#include "../ctype.h"
 
 int tolower(int c)
 {
